fix: resolve clippy warnings across workspace
- Implement FromStr trait for Dialect instead of custom from_str method
- Collapse nested if-let statements into let-chains (upgrade.rs)
- Use let destructuring for infallible match (generate.rs)
- Remove unnecessary u64 cast (status.rs)

+++ cli/src/commands/status.rs
@@ -79,7 +79,7 @@ fn format_timestamp(ms: u64) -> String {
     }
 
     let secs = ms / 1000;
-    let datetime = std::time::UNIX_EPOCH + std::time::Duration::from_secs(secs as u64);
+    let datetime = std::time::UNIX_EPOCH + std::time::Duration::from_secs(secs);
 
     // Format as ISO date
     format!("{:?}", datetime)
